converting OK-switch
