+++ hardware/switches/osw12.py
@@ -16,6 +16,7 @@ along with Qudi. If not, see <http://www.gnu.org/licenses/>.
 Copyright (c) the Qudi Developers. See the COPYRIGHT.txt file at the
 top-level directory of this distribution and at <https://github.com/Ulm-IQO/qudi/>
 """
+# TODO: convert interface
 
 import visa
 from core.module import Base
